esnet_gui_deployment/current/spin-docker-netpredict/requirements.txt
click==8.0.1
cycler==0.10.0
decorator==4.4.2
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.5.1
networkx==2.5.1
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
Werkzeug==2.0.1
mysql-connector-python==8.0.27
# SQLAlchemy==1.4.29
# PyMySQL==1.0.2 

